feat: add files for individual rules and documentation for each section (#4)
Add files for individual rules and documentation around each section of
the Google Python Style Guide.

This PR contains the following changes:

* GPSG rules were reordered to match the one in the new version of the
public documentation (still under development)
* each section of the GPSG rules got their own summary in the
`.sourcery.yaml` file
* the folder `examples/individual_rules` was created containing one file
per section of rules. Those files are meant to be populated with
examples in a follow-up PR, and already contain a summary docstring

Please make sure that you agree with the new documentation comments and
docstrings, as well as with the overall structure of the project.

Note that most of the links do not work since they refer to the new
version of the public docs, which has not been released yet. The issue
https://github.com/sourcery-ai/google-python-style-guide-demo/issues/3
will keep track of that.